refactor: 重构组件架构，实现自给自足的组件设计
## 主要改进

### 组件拆分与优化
- 将 FilesPageComponent 拆分为 FilesHeader、FilesSidebar、FilesArea 三个独立组件
- 实现组件自给自足原则，大幅减少 props 传递：
  - FilesHeader: 6 props → 0 props
  - FilesSidebar: 1 prop → 0 props
  - FilesArea: 6 props → 1 prop (currentPath)

### 统一状态管理
- 创建 DirectoryContext 统一管理目录选择状态
- 移除 UploadContext 中重复的 directoryId 状态
- 实现单一数据源原则，消除状态同步问题

### Context 架构优化
- 智能的 Provider 层级设计，延迟初始化避免无效状态
- 自动从 useAuth 获取默认目录，无需手动配置
- 上传完成回调内置化，自动刷新文件列表和存储空间

### 类型安全增强
- selectedDirectoryId 改为非可选类型，提升类型安全性
- 移除不必要的空值检查和防御性编程
- 简化组件验证逻辑，移除冗余的目录ID验证

### 性能优化
- 构建产物减小 0.6 kB，移除不必要的依赖和代码
- 组件职责明确，减少重复渲染
- Context 作用域合理化，避免全局状态污染
