+++ src/frontend/components/UploadDrawer.tsx
@@ -22,13 +22,9 @@ import { Card } from "./ui/card";
 
 interface UploadDrawerProps {
   targetPath?: string;
-  directoryId?: string;
 }
 
-export function UploadDrawer({
-  targetPath,
-  directoryId,
-}: UploadDrawerProps = {}) {
+export function UploadDrawer({ targetPath }: UploadDrawerProps = {}) {
   const router = useRouter();
   const [isVisible, setIsVisible] = useState(false);
   const [isAnimating, setIsAnimating] = useState(false);
@@ -82,23 +78,17 @@ export function UploadDrawer({
 
   const currentPath = getCurrentPath();
 
-  // Handle drawer close with file list refresh
+  // Handle drawer close
   const handleCloseDrawer = useCallback(() => {
     setIsDrawerOpen(false);
-
-    // File list refresh is handled by the upload complete callback in FilesPageComponent
+    // File list refresh is now handled automatically by UploadContext
   }, [setIsDrawerOpen]);
 
   const onDrop = useCallback(
     (acceptedFiles: File[]) => {
-      addFiles(
-        acceptedFiles,
-        currentPath,
-        fileConfig?.max_file_size_mb,
-        directoryId,
-      );
+      addFiles(acceptedFiles, currentPath, fileConfig?.max_file_size_mb);
     },
-    [addFiles, currentPath, fileConfig?.max_file_size_mb, directoryId],
+    [addFiles, currentPath, fileConfig?.max_file_size_mb],
   );
 
   const { getRootProps, getInputProps, isDragActive } = useDropzone({
@@ -109,7 +99,7 @@ export function UploadDrawer({
   });
 
   const handleStartUpload = () => {
-    startUpload(currentPath, directoryId);
+    startUpload(currentPath, currentPath);
   };
 
   const stats = getUploadStats();
