+++ src/frontend/contexts/DirectoryContextDefinition.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+export interface DirectoryContextType {
+  selectedDirectoryId: string;
+  setSelectedDirectoryId: (directoryId: string) => void;
+}
+
+export const DirectoryContext = createContext<DirectoryContextType | undefined>(
+  undefined,
+);
