+++ src/frontend/routes/__root.tsx
@@ -2,10 +2,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { Outlet, createRootRoute } from "@tanstack/react-router";
 import { Toaster } from "sonner";
-import { UploadDrawer } from "../components/UploadDrawer";
-import { UploadIndicator } from "../components/UploadIndicator";
 import { AuthProvider } from "../contexts/AuthContext";
-import { UploadProvider } from "../contexts/UploadContext";
 import "../styles/globals.css";
 
 // 创建 Query Client 实例
@@ -26,19 +23,13 @@ export const Route = createRootRoute({
   component: () => (
     <QueryClientProvider client={queryClient}>
       <AuthProvider>
-        <UploadProvider>
-          <div className="min-h-screen bg-background">
-            <Outlet />
-            <ReactQueryDevtools initialIsOpen={false} />
+        <div className="min-h-screen bg-background">
+          <Outlet />
+          <ReactQueryDevtools initialIsOpen={false} />
 
-            {/* Global upload components */}
-            <UploadDrawer />
-            <UploadIndicator />
-
-            {/* Toast notifications */}
-            <Toaster richColors position="top-right" />
-          </div>
-        </UploadProvider>
+          {/* Toast notifications */}
+          <Toaster richColors position="top-right" />
+        </div>
       </AuthProvider>
     </QueryClientProvider>
   ),
